src/test/java/ru/yandex/practicum/filmorate/storage/impl/db/GenreDbStorageTest.java
package ru.yandex.practicum.filmorate.storage.impl.db;

import lombok.RequiredArgsConstructor;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.DirtiesContext;
import ru.yandex.practicum.filmorate.model.Genre;

import java.util.List;
import java.util.Optional;

import static org.assertj.core.api.Assertions.assertThat;

@SpringBootTest
@AutoConfigureTestDatabase
@RequiredArgsConstructor(onConstructor_ = @Autowired)
@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
class GenreDbStorageTest {
    @Autowired
    private GenreDbStorage genreStorage;

    @Test
    void findAll() {
        List<Genre> genreList = genreStorage.findAll();
        assertThat(genreList.size()).isEqualTo(6);
        assertThat(genreList.get(4).getName()).isEqualTo("Документальный");
    }

    @Test
    void findGenreById() {
        Optional<Genre> genreOptional = Optional.ofNullable(genreStorage.findGenreById(2));
        assertThat(genreOptional)
                .isPresent()
                .hasValueSatisfying(genre ->
                        assertThat(genre).hasFieldOrPropertyWithValue("id", 2)
                                .hasFieldOrPropertyWithValue("name", "Драма")
                );
    }
}